# api/swagger/models.gen.go
// Package swagger provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package swagger

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message *string `json:"message,omitempty"`
}

// IDResponse defines model for IDResponse.
type IDResponse struct {
	Id *string `json:"id,omitempty"`
}

// ProductRequest defines model for ProductRequest.
type ProductRequest struct {
	Id    *int64  `json:"id,omitempty"`
	Name  string  `json:"name"`
	Price float32 `json:"price"`
}

// TelegramBot defines model for TelegramBot.
type TelegramBot struct {
	Description *string `json:"description,omitempty"`
	Id          *string `json:"id,omitempty"`
	ImageUrl    *string `json:"image_url,omitempty"`
	Name        *string `json:"name,omitempty"`
	Token       *string `json:"token,omitempty"`
}

// TelegramBotRequest defines model for TelegramBotRequest.
type TelegramBotRequest struct {
	Description *string `json:"description,omitempty"`
	ImageUrl    *string `json:"image_url,omitempty"`
	Name        string  `json:"name"`
	Token       string  `json:"token"`
}

// DeleteTelegramBotParams defines parameters for DeleteTelegramBot.
type DeleteTelegramBotParams struct {
	// Id ID of the telegram bot to delete
	Id string `form:"id" json:"id"`
}

// CreateProductJSONRequestBody defines body for CreateProduct for application/json ContentType.
type CreateProductJSONRequestBody = ProductRequest

// CreateTelegramBotJSONRequestBody defines body for CreateTelegramBot for application/json ContentType.
type CreateTelegramBotJSONRequestBody = TelegramBotRequest